requirements.txt
gitpython
scikit-learn
numpy